working datamodel
